configure: some changes to config.h, CSRG, and XTOOLLIB impl
We create autotools_config.h instead of config.h to make it clear  and
avoid collisions.

We use a different method to define CSRG_BASED on BSD systems, and fix
a bug in definition of XTOOLLIB.

XTOOLLIB should probably be checked out - we should get some info from
AC_PATH_X and AC_PATH_XTRA instead of hardcoding it.  To be looked at
later.