DtSearch: get it to build, fix up some confusion
There are/were two dproto.h files, one in DtSearch, one in raima.  The
one in raima is the one we want, so replace the global Dt/dproto.h
with it.  Also correct a prototype declaration error in that file.

Delete the dproto.h file in DtSearch, it's garbage and unused.

Next, redo the way yacc is used.  Autotools does "the right thing" by
simply including the .y file in *_SOURCES and using
BUILT_SOURCES/CLEANFILES to manage it.  No need for special rules to
handle yacc and lex.