configure: add programs/ dthelp, dsdm, dtmail, dtpad, and dtfile
At this stage, these certainly won't actually build yet.

Just fix up the relevent Makefile.am files so that autogen does not
emit errors and warnings for them.

Removed AIX/HPUX support in Makefile.am files.  No point in
propogating that stuff when we've already removed much of that
unmaintained code from the codebase.

Commented out all of the Sun Pro stuff.  Someone whos using that will
need to go through and fix it.  This is mostly in dtmail and dthelp.

In fact, someone who does Solaris in general will need to go through
this stuff.

Next up, we'll replace any remaining 'if SUN' conditionals with 'if
SOLARIS' which is a more appropriate name and was already defined in
configure.ac.

Then we'll see about getting these new directories building.