DtSvc: we can't use archive .a libs in building libDtSvc.so
We were building subsets of this library as archive.a objects, then
trying to combine them into a shared object.  This warnings, and link
failures due to the lack of 0fPIC when building the archive libraries.

Now, we do not buld archive libraries, and the main libDtSvc.so
library is build by adding all of the .lo file in the component
subdirectories.

In a future commit, we could just remove the SUBDIR builds (DtUtil1,
etc) and build the sub objects directly.