configure: Several changes related to locating cpp and ksh
For ksh, we need a full pathname.  AC_CHECK_PROGS only sets the name,
so we can't use that (think of a "#!" in a shell script.

We use some shell scripting to locate the ksh pathname.

While on that subject, the current use of CPP (gcc -E) as a general
preprocessor does not work very well.  I messes up whitespace,
adds/translates random whitespace, and complains bitterly about single
quotes (') in various places like comments.  It's not usable for what
CDE needs.

So, now we use GENCPP.  Using shell scripting like that used for ksh,
we locate the cpp program, and set GENCPP to "/full/path/to/cpp
-traditional -nostdinc".  This is what Linux uses now in an Imake
build, and it works fine.  We'll have to see what the BSD/Solari do.

We might need to just include BSD's "tradcpp" into the build and use
that.  It too works well in limited testing, but eats blank lines.  We
can live with that if we have to.