configure: rework libXinerama detection a bit, remove hardcoded -lXinerama's
Make -lXinerama a dep on libDtXinerama so it doen't need to be
specified in Makefile.am files or in LIBS as it was being done
previously.

This still needs a little work, ideally Xinerama should be completely
optional and only enabled if present.  But we can save that for
later.