configure: set a compiler search option
... and move the PROG_CC and PROG_CXX checks toward the beginning.
Without the move the compiler search arguments are ignored.  We search
for 'cc gcc clang' and 'c++ g++ clang++', in that order for CC and CXX
respectively.

This means there should be no need to override the CC
and CXX variables on the configure command line for BSD systems.

On OpenBSD for example, cc links to clang.  On Linux, cc links to
gcc.  You can still override these if you like with CC=... and
CXX=... on the configure command line.

gmake (MAKE=gmake) still needs to be used on the BSD's though.

With these and previous changes, CDE builds, installs, and runs
correctly on OpenBSD 6.5 now.