FreeBSD.cf,DragonFly.cf: disable PAM support by default
These system can support PAM, but it requires a port of the NetBSD
module to do so.  As a result, this support is disabled by default.

See https://bugs.freebsd.org/bugzilla/show_bug.cgi?id=253073 for a
patch to add this support to the OS.  Then you can set HasPamLibrary
to YES in the respective .cf files, rebuild CDE, and try it out.