utempter: complete the implementation
Liang Chang added utempter support to the dtterm widget in
pre-autoconf CDE.  While the code itself was merged, it was still not
"turned on" for autoconf builds.

This commit completes the implementation allowing dtterm on Linux and
the BSDs to be installed without having to be setuid root -- as long
as the libutempter headers and libraries are installed.