dtappbuilder: try to fix a dep error causeing every make to rebuild
With the addition of dtappbuilder to autotools, every make run
(including make install) causes src/ab to regenerate files that cause
the whole directory to be recompiled, even on a 'make install'.

This seems to be primarily caused by dtbuilder.msg (and possibly
dtbuilder.c) being modified after it is generated, which is apparently
normal behavior for dtcodegen.

This fix sets both dtbuilder.msg and dtbuilder.c as "order-dependant
prequisites" so that they are always generated first (there are
already rules to do generate them) and then subsequently the only
thing required is that they exist since we don't care if they are
modified after that point.  If you need to make changes to them, edit
their respective .src files instead.

This seems to resolve the issue, but perhaps at the expense of adding
another requirement to use gmake.

We also need to disable parallel building in this module :(