dtcm: timeops.c/getdate.y: fix fbsd build errors
Commit e0508b31 introduced build errors on FreeBSD.  This corrects
them.

HAVE_DECL_TIMEZONE should only be used to determine whether or not the
'timezone' variable is defined in a header file or whether it must be
specifically 'extern'ed.

On fbsd, a definition exists, but it is a function in libc and not an
integer timezone value that can be mutliplied or divided.