dthelp.dt.src: hack to make Dthelpview ACTION work
For reasons I have been unable to determine, running:

dtaction Dthelpview

or Selecting 'Help Manager' from the front panel fails trying to run
the appropriate EXEC_STRING.  It get's corrupted in some fashion and
you will usually see an error in ~/.dt/errorlog of the form:

/usr/dt/bin/t/bin/dthelpview: No such file or directory

or more commonly:

/bin/sh: /us: No such file or directory

It is not clear where this corruption is taking place, but it shows up
in 'dttypes' output as well.

This hack simply removes 10 spaces between the two commands that are
run:

 EXEC_STRING     /bin/sh -c '/usr/dt/bin/dthelpgen -dir $HOME/.dt/help/$DTUSERSESSION; \
                           /usr/dt/bin/dthelpview -helpVolume browser
                           '

Which seems to solve (hide) the problem.  Finding the real cause of
this will be saved for a later date when more is known and in
particular whether this happens with other ACTIONS.